feat(#383): refactoring types in relation to UPLC
