+++ symbolic-uplc/src/ZkFold/Symbolic/UPLC/Fun.hs
@@ -1,0 +1,39 @@
+{-# LANGUAGE AllowAmbiguousTypes    #-}
+{-# LANGUAGE ConstraintKinds        #-}
+{-# LANGUAGE DataKinds              #-}
+{-# LANGUAGE FlexibleContexts       #-}
+{-# LANGUAGE FlexibleInstances      #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE GADTs                  #-}
+{-# LANGUAGE KindSignatures         #-}
+{-# LANGUAGE RankNTypes             #-}
+{-# LANGUAGE ScopedTypeVariables    #-}
+{-# LANGUAGE TypeApplications       #-}
+{-# LANGUAGE TypeOperators          #-}
+{-# LANGUAGE UndecidableInstances   #-}
+
+
+module ZkFold.Symbolic.UPLC.Fun (Fun (..)) where
+
+import           Data.Function                      ((.))
+
+import           ZkFold.Algebra.Class               (FromConstant (..))
+import qualified ZkFold.Symbolic.Data.Maybe         as Symbolic
+import           ZkFold.Symbolic.UPLC.Class
+import           ZkFold.UPLC.BuiltinType
+
+
+-- | Symbolic function of a definite UPLC signature.
+data Fun (s :: [BuiltinType]) (t :: BuiltinType) c where
+  -- | Fully applied (saturated) function.
+  FSat :: IsData t v c => Symbolic.Maybe c v -> Fun '[] t c
+  -- | A function which returns another (possibly saturated) function.
+  FLam :: IsData s v c => (v -> Fun ss t c) -> Fun (s ': ss) t c
+
+instance IsData t v c => FromConstant (Symbolic.Maybe c v) (Fun '[] t c) where
+  fromConstant = FSat
+
+instance
+    (IsData s v c, FromConstant f (Fun ss t c))
+    => FromConstant (v -> f) (Fun (s ': ss) t c) where
+  fromConstant f = FLam (fromConstant . f)
